split mlfsm and qchem into seperate modules, add sella module
